moving types and styles to their expected files
